+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "@/App.tsx";
-import { UserContextProvider } from "./firebase/UserContext";
+import { UserContextProvider } from "@/contexts/UserContext";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
